+++ CHANGELOG.md
@@ -7,8 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Upcoming
 
+## [0.4.8] 2020-03-26
+### Added
+- Optional Environment variable AMQP_PREFETCH_COUNT
 
-## [0.4.0] - ?
+### Changed
+- Renamed env var REDIS_URI to REDIS_CONFIGURATION_STRING (same functionality)
+
+
+## [0.4.0] - 2020-03-04
 ### Added
 - CI